fix: address all PR review comments
- Preserve null/undefined behavior: use `value == null` in batch
  validation and XML generation to reject both null and undefined
- Move XMLParseError to src/errors.js with all other error classes
- Revert rejectUnauthorized default to false (Niagara uses self-signed certs)
- Extract xmlElementForValue to src/helpers.js (was duplicated in
  standard.js and batch.js), add tests
- Use startsWith for history preset matching (Niagara returns names
  like "yearToDate (limit=1000)")
- Bump CI matrix to Node 18/20/22 and engine requirement to >=18
  (Node 14 is EOL, can't parse modern lockfiles)
