+++ CLAUDE.md
@@ -26,7 +26,7 @@ Exports `ObixInstance` and `BQLInstance`. Both validate constructor params (host
 ### Source (`src/`)
 
 - **`errors.js`** — All custom error classes consolidated here. `HTTPError` and `BQLHTTPError` extend a shared `BaseHTTPError` that matches errors by HTTP status code first, then falls back to message string matching. All errors expose `friendlyError` and `inDepthError` properties. Also contains `ProtocolError`, `PathError`, `PathTraversalError`, `InvalidTypeError`, `UnknownTypeError`, and the history/BQL query errors.
-- **`axios.js`** — Creates axios instances. oBIX instances use XML-to-JSON transform (via `xml-js`) and throw `XMLParseError` on parse failure. BQL instances return raw HTML. Both handle auth, timeouts (default 10s), cookie persistence, and response error interception. `rejectUnauthorized` defaults to `true`; pass `false` for self-signed certs.
+- **`axios.js`** — Creates axios instances. oBIX instances use XML-to-JSON transform (via `xml-js`) and throw `XMLParseError` on parse failure. BQL instances return raw HTML. Both handle auth, timeouts (default 10s), cookie persistence, and response error interception. `rejectUnauthorized` defaults to `false` (Niagara controllers commonly use self-signed certs); pass `true` to enforce certificate validation.
 - **`helpers.js`** — `stripPaths` normalizes paths and rejects `..` traversal. `replaceSpecialChars` coerces all values to string before XML-escaping. `makeArray` ensures array type.
 - **`parsers/values.js`** — Parses oBIX XML-to-JSON responses into `{ path, value, action }` objects. Handles `real`, `bool`, `str`, and `enum` types. `buildOutputList` guards against missing `data.list`.
 - **`parsers/errors.js`** — Parses oBIX error responses (`err` elements) and throws appropriate error classes.
