+++ package.json
@@ -23,7 +23,7 @@
   "author": "Adam Zimmerman",
   "license": "ISC",
   "engines": {
-    "node": ">=14"
+    "node": ">=18"
   },
   "bugs": {
     "url": "https://github.com/Rise-Building-Technology/obix-js/issues"
